Update deployment info (#1707)
