+++ pkg/deployments/action-ids/arbitrum/action-ids.json
@@ -348,14 +348,5 @@
         "disable()": "0xe59ccea4d4a98c33b7c0b3c05a3d65ff930273fa5ee03855a3104befc6d1d362"
       }
     }
-  },
-  "20220817-composable-stable-pool": {
-    "ComposableStablePoolFactory": {
-      "useAdaptor": false,
-      "actionIds": {
-        "create(string,string,address[],uint256,address[],uint256[],bool[],uint256,address)": "0xd6533f30a3d63e24d5a4aecaa0b60857fde5808b210bb73c5af99332312d5d91",
-        "disable()": "0xf33f1f71e25eef86857e743ed24308664c252202a830439e98b6b22338a05784"
-      }
-    }
   }
 }
