+++ pkg/deployments/action-ids/optimism/action-ids.json
@@ -238,14 +238,5 @@
         "disable()": "0xe55819ea19bc69b0b5b19180d8847296d579facecfbe27ba16d7d9940aca0ae5"
       }
     }
-  },
-  "20220817-composable-stable-pool": {
-    "ComposableStablePoolFactory": {
-      "useAdaptor": false,
-      "actionIds": {
-        "create(string,string,address[],uint256,address[],uint256[],bool[],uint256,address)": "0xb878bca15e4502b73f0316b18d4be1ce87b80662e185d34de67f9434870e44c3",
-        "disable()": "0xade556e97ab1839aa7cff8fea1423e2e5c347eeb2c42b771cd4a56b9fc254cfd"
-      }
-    }
   }
 }
