--- pkg/deployments/tasks/20220817-composable-stable-pool/index.ts
@@ -1,10 +1,0 @@
-import Task from '../../src/task';
-import { TaskRunOptions } from '../../src/types';
-import { ComposableStablePoolDeployment } from './input';
-
-export default async (task: Task, { force, from }: TaskRunOptions = {}): Promise<void> => {
-  const input = task.input() as ComposableStablePoolDeployment;
-
-  const args = [input.Vault, input.ProtocolFeePercentagesProvider];
-  await task.deployAndVerify('ComposableStablePoolFactory', args, from, force);
-};
