--- pkg/deployments/tasks/20220817-composable-stable-pool/output/arbitrum.json
@@ -1,4 +1,0 @@
-{
-  "ComposableStablePoolFactory": "0x44A41fE82c9539C1aadaBEE5B9C3f327eba5C8a3",
-  "timestamp": 1660755513763
-}
