--- pkg/deployments/tasks/20220817-composable-stable-pool/output/goerli.json
@@ -1,4 +1,0 @@
-{
-  "ComposableStablePoolFactory": "0x239e55F427D44C3cc793f49bFB507ebe76638a2b",
-  "timestamp": 1660754799749
-}
