--- pkg/deployments/tasks/20220817-composable-stable-pool/output/mainnet.json
@@ -1,4 +1,0 @@
-{
-  "ComposableStablePoolFactory": "0xDc6d62ae091ea77331542042A72e5E38B188837C",
-  "timestamp": 1660755681344
-}
