--- pkg/deployments/tasks/20220817-composable-stable-pool/output/optimism.json
@@ -1,4 +1,0 @@
-{
-  "ComposableStablePoolFactory": "0x7C71313d3b84728cA0FD4cDaD8c86067E3548dCe",
-  "timestamp": 1660755421408
-}
