--- pkg/deployments/tasks/20220817-composable-stable-pool/output/polygon.json
@@ -1,4 +1,0 @@
-{
-  "ComposableStablePoolFactory": "0x967F7AdD4Fd5AF0553B7A45F225ec26EDD699E61",
-  "timestamp": 1660755000494
-}
