--- pkg/deployments/tasks/20220817-composable-stable-pool/readme.md
@@ -1,12 +1,0 @@
-# 2022-08-17 - Composable Stable Pool
-
-Deployment of `ComposableStablePoolFactory`, which allows creating Stable Pools that are suitable to be included in other Pools.
-
-## Useful Files
-
-- [Ethereum addresses](./output/mainnet.json)
-- [Polygon addresses](./output/polygon.json)
-- [Arbitrum addresses](./output/arbitrum.json)
-- [Optimism addresses](./output/optimism.json)
-- [`ComposableStablePool` ABI](./abi/ComposableStablePool.json)
-- [`ComposableStablePoolFactory` ABI](./abi/ComposableStablePoolFactory.json)
